Simplify daml-react-hooks tests
... and reflect the circumstance that they actually test the `LedgerStore` in
the file name.
